fix: accept urlToPath callback for people without URL.pathname polyfill
RN's default URL polyfill might not provide `.pathname`, which we use to extract file extnensions.
Instead of requiring everyone to use a polyfill in their app, we instead now take an optional
`urlToPath` callback which is used to extract file extensions. If it's not provided, then files
will be saved without extensions.

BREAKING CHANGE: The default behavior is now changed to not save files with extensions. To make this
change not-breaking would have required that this package accept and use a polyfill for URL, which
feels like it could cause greater problems for people who already use other polyfills.

fix #23